fix(2021): day 19 cleanup
